refactor: Main 컴포넌트에서 초대 관련 훅 정리
